refactor: 알림 한개짜리 컴포넌트 분리
