+++ components/commons/myNotifications/Notifications.tsx
@@ -1,9 +1,8 @@
 'use client';
-import { useDeleteNotification } from '@/apis/my-notifications/useDeleteMyNotification';
 import useGetMyNotifications from '@/apis/my-notifications/useGetMyNotifications';
-import formatDateDiff from '@/utils/formatDateDiff';
 import Image from 'next/image';
 import { useCallback, useEffect, useRef } from 'react';
+import NotificationItem from './NotificationsItem';
 
 interface MyNotificationsProps {
   onClose: () => void;
@@ -15,12 +14,7 @@ const Notifications = ({
   notificationIconRef,
 }: MyNotificationsProps) => {
   const { data, error, isLoading } = useGetMyNotifications();
-  const { mutate } = useDeleteNotification();
-  const notificationsRef = useRef<HTMLDivElement | null>(null);
-
-  const handleDeleteNotification = (notificationId: number) => {
-    mutate({ notificationId });
-  };
+    const notificationsRef = useRef<HTMLDivElement | null>(null);
 
   const handleClickOutside = useCallback(
     (event: MouseEvent) => {
@@ -42,29 +36,6 @@ const Notifications = ({
       document.removeEventListener('mousedown', handleClickOutside);
     };
   }, [handleClickOutside]);
-
-  const renderContent = (content: string) => {
-    if (!content) return null;
-
-    const parts = content.split(/(승인|거절)/g);
-    return parts.map((part, index) => {
-      if (part === '승인') {
-        return (
-          <span key={index} className="text-blue300">
-            승인
-          </span>
-        );
-      }
-      if (part === '거절') {
-        return (
-          <span key={index} className="text-red100">
-            거절
-          </span>
-        );
-      }
-      return part;
-    });
-  };
 
   return (
     <div
@@ -90,37 +61,12 @@ const Notifications = ({
       {data && data.notifications.length > 0 && (
         <div className="flex flex-col gap-2 mt-4 h-[632px] overflow-scroll scrollbar-hide mobile:h-full">
           {data.notifications.map((item) => (
-            <div
+            <NotificationItem
               key={item.id}
-              className="h-[120px] px-3 py-4 rounded-[5px] border border-[#CBC9CF] bg-white"
-            >
-              <div className="flex justify-between items-start">
-                <div
-                  className={`${
-                    item.content.includes('승인') ? 'bg-blue300' : 'bg-red100'
-                  } w-[5px] h-[5px] mt-1 rounded-full`}
-                ></div>
-                <Image
-                  src="/icons/btn-X-medium.svg"
-                  alt="알림 삭제 버튼"
-                  width={24}
-                  height={24}
-                  className="cursor-pointer"
-                  onClick={() => handleDeleteNotification(item.id)}
-                />
-              </div>
-              <div className="flex flex-col justify-between h-[62px]">
-                <p
-                  className="mb-1 text-body2-regular"
-                  style={{ wordBreak: 'keep-all', wordWrap: 'break-word' }}
-                >
-                  {renderContent(item.content)}
-                </p>
-                <p className="text-caption text-gray400">
-                  {formatDateDiff(item.updatedAt)}
-                </p>
-              </div>
-            </div>
+              id={item.id}
+              content={item.content}
+              updatedAt={item.updatedAt}
+            />
           ))}
         </div>
       )}
